requirements.txt
Flask==1.1.1
pymongo==3.8.0
requests==2.18.4
python_dateutil==2.8.0
fbchat==1.7.2
gunicorn==19.9.0
schedule==0.6.0

